feat: mysql_spider查询权限_权限规则查询提高效率_dbm跨业务迁移权限规则 #5583

+++ dbm-services/mysql/db-priv/handler/admin_password.go
@@ -33,7 +33,7 @@ func (m *PrivService) GetPassword(c *gin.Context) {
 		return
 	}
 	SendResponse(c, err, ListResponse{
-		Count: int64(count),
+		Count: count,
 		Items: batch,
 	})
 	return
@@ -103,7 +103,7 @@ func (m *PrivService) GetMysqlAdminPassword(c *gin.Context) {
 		return
 	}
 	SendResponse(c, err, ListResponse{
-		Count: int64(count),
+		Count: count,
 		Items: batch,
 	})
 	return
